Added naive versions for the additive and affine couplings, splitted unittests
For now disabled the affine coupling from the unittests
